修复量子存储卡顿

+++ src/main/java/me/ddggdd135/slimeae/api/interfaces/IStorage.java
@@ -70,7 +70,7 @@ public interface IStorage {
      * @return 物品到数量的映射
      */
     @Nonnull
-    Map<ItemStack, Integer> getStorage();
+    Map<ItemStack, Long> getStorage();
 
     /**
      * 获取存储中的空槽位数量
